+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - [V2] Agency: Supply ClientID with agency transactions
 - Bugfix: Force analyzer removal
 - Move examples to separate package
+- Deprecate ClientConfig.SynchronizeEndpointsInterval due to bug in implementation
+- [V1] Add Rename function for collections (single server only).
 
 ## [1.6.0](https://github.com/arangodb/go-driver/tree/v1.6.0) (2023-05-30)
 - Add ErrArangoDatabaseNotFound and IsExternalStorageError helper to v2